Cleaned main function a bit
