+++ .vscode/settings.json
@@ -1,10 +1,14 @@
 {
     "files.associations": {
-        "ostream": "cpp",
-        "iosfwd": "cpp",
-        "iostream": "cpp",
-        "fstream": "cpp",
-        "new": "cpp",
-        "*.tcc": "cpp"
-    }
+		"ostream": "cpp",
+		"iosfwd": "cpp",
+		"iostream": "cpp",
+		"fstream": "cpp",
+		"new": "cpp",
+		"*.tcc": "cpp",
+		"stdexcept": "cpp",
+		"cstring": "cpp",
+		"vector": "cpp",
+		"string": "cpp"
+	}
 }
