+++ Makefile
@@ -7,10 +7,19 @@ RESPONSE_DIR = ./response/
 RESPONSE_FILES = Response.cpp
 
 SRC_DIR = ./
-SRC_FILES = main.cpp socket.cpp utils.cpp
+SRC_FILES = main.cpp utils.cpp
+
+SOCKET_DIR = ./sockets/
+SOCKET_FILES = socket.cpp
+
+EPOLL_DIR = ./epoll/
+EPOLL_FILES = epoll.cpp
+
+EVENT_DIR = ./events/
+EVENT_FILES = eventLoop.cpp
 
 PARSE_DIR = ./parsing/
-PARSE_FILES = ServerConfig.cpp LocationConfig.cpp parseConfFile.cpp ConfigValidator.cpp
+PARSE_FILES = ServerConfig.cpp LocationConfig.cpp parseConfFile.cpp ConfigValidator.cpp config.cpp
 
 CGI_DIR = ./cgi/
 CGI_FILES = cgi_request.cpp
@@ -18,13 +27,16 @@ CGI_FILES = cgi_request.cpp
 SRC += $(addprefix $(SRC_DIR), $(SRC_FILES))
 SRC += $(addprefix $(REQUEST_DIR), $(REQUEST_FILES))
 SRC += $(addprefix $(RESPONSE_DIR), $(RESPONSE_FILES))
+SRC += $(addprefix $(SOCKET_DIR), $(SOCKET_FILES))
+SRC += $(addprefix $(EPOLL_DIR), $(EPOLL_FILES))
+SRC += $(addprefix $(EVENT_DIR), $(EVENT_FILES))
 SRC += $(addprefix $(PARSE_DIR), $(PARSE_FILES))
 SRC += $(addprefix $(CGI_DIR), $(CGI_FILES))
 
 OBJ = $(SRC:.cpp=.o)
 COMPILER = c++ -std=c++17
 CFLAGS = -Wall -Wextra -Werror
-LFLAGS = 
+LFLAGS =
 
 all: $(NAME)
